Full Stack with LLM/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
gspread>=6.0.0
google-auth>=2.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0

# For local model inference (optional, ~4GB+ download)
# transformers>=4.40.0
# torch>=2.2.0
# accelerate>=0.29.0
